Through the styles an override can be provided for specified road classes such that traffic segments with an unknown traffic congestion value will be given the traffic color for low traffic rather than unknown.

+++ libnavigation-ui/src/main/java/com/mapbox/navigation/ui/route/NavigationMapRoute.java
@@ -433,7 +433,7 @@ public class NavigationMapRoute implements LifecycleObserver {
             ? Collections.emptyList() : routeStyleDescriptors;
     final RouteLayerProvider layerProvider = getLayerProvider(routeStyleDescriptorsToUse, context, styleRes);
 
-    final float vanishingPointOffset = routeLine.getVanishPointOffset();
+    final double vanishingPointOffset = routeLine.getVanishPointOffset();
     routeLine = new MapRouteLine(
         context,
         style,
